+++ charset_normalizer/api.py
@@ -304,7 +304,7 @@ def from_bytes(
 
                 md_ratios.append(
                     mess_ratio(
-                        chunk, threshold, explain is True and len(cp_isolation) <= 2
+                        chunk, threshold, explain is True and cp_isolation and len(cp_isolation) <= 2
                     )
                 )