+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [3.0.0b3](https://github.com/Ousret/charset_normalizer/compare/3.0.0b2...3.0.0b3) (2022-10-??)
 
 ### Added
-- Extend the capability of explain=True when cp_isolation contain at most two entries, will log in details the Mess-detector results
+- Extend the capability of explain=True when cp_isolation contains at most two entries (min one), will log in details of the Mess-detector results
 
 ### Changed
 - Build with static metadata using 'build' frontend