+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 - Build with static metadata using 'build' frontend
+- Make the language detection stricter
 
 ### Fixed
 - CLI with opt --normalize fail when using full path for files